requirements.txt
ads==0.12.3
aniso8601==4.1.0
APScheduler==3.6.0
asn1crypto==0.24.0
astroid==2.1.0
astropy==3.2.1
astropy-helpers==3.2.1
astroquery==0.3.9
atomicwrites==1.3.0
attrs==18.2.0
autopep8==1.4.3
beautifulsoup4==4.7.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorhash==1.0.2
configparser==3.7.1
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
entrypoints==0.3
flake8==3.7.5
flake8-mypy==17.8.0
Flask==1.0.2
Flask-MonitoringDashboard==2.1.3
flask-restplus==0.12.1
GeoAlchemy2==0.6.3
gunicorn==19.9.0
html5lib==1.0.1
httpretty==0.8.10
idna==2.8
inflect==2.1.0
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==1.1.0
jeepney==0.4
Jinja2==2.10.1
jsonschema==2.6.0
jupyter-core==4.4.0
keyring==19.0.2
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
matplotlib==3.1.0
mccabe==0.6.1
mock==3.0.5
more-itertools==7.0.0
mypy==0.670
mypy-extensions==0.4.1
nbformat==4.4.0
numpy==1.16.2
passlib==1.7.1
pg8000==1.13.1
plotly==3.7.1
pluggy==0.11.0
psutil==5.6.1
psycopg2-binary==2.8.3
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.0
pylint==2.2.2
PyMySQL==0.9.3
pyparsing==2.4.0
pytest==4.5.0
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2018.9
redis==3.3.8
requests>=2.20.0
retrying==1.3.3
rq==1.1.0
scipy==1.3.0
SecretStorage==3.1.1
six==1.12.0
soupsieve==1.9.1
sqlacodegen==2.0.1
SQLAlchemy==1.3.5
sqlalchemy-stubs==0.1
sseclient==0.0.24
synphot==0.1.3
traitlets==4.3.2
typed-ast==1.3.1
typing-extensions==3.7.2
tzlocal==1.5.1
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
werkzeug>=0.15.3
wrapt==1.11.1
git+git://github.com/Small-Bodies-Node/catch.git@v0.3.2#egg=catch
git+git://github.com/NASA-Planetary-Science/sbpy.git@v0.1.1#egg=sbpy
git+git://github.com/Small-Bodies-Node/sbsearch.git@v1.0.4#egg=sbsearch
